requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.5
cryptography==3.4.8
decorator==5.1.1
distro==1.7.0
Django==4.2.1
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
executing==1.2.0
factory-boy==3.2.1
Faker==18.9.0
fasteners==0.14.1
future==0.18.2
httplib2==0.20.2
idna==2.10
importlib-metadata==4.6.4
inflection==0.5.1
ipython==8.13.2
itypes==1.2.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
packaging==23.1
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.38
protobuf==3.12.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
ssh-import-id==5.11
stack-data==0.6.2
traitlets==5.9.0
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.6
xdg==5.0.0
zipp==1.0.0
